+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-input/example/DefaultDateInputField.jsx
@@ -1,24 +1,46 @@
-import React, { useState } from 'react';
+import React, { useState, useRef } from 'react';
 
 import DateInputField from 'terra-date-input/lib/DateInputField';
 
 const Example = () => {
+  const ref = useRef();
   const [value, setValue] = useState('');
   const [isInvalid, setIsInvalid] = useState(false);
+  const [errorMessage, setErrorMessage] = useState('');
+
+  const handleClick = () => {
+    setIsInvalid(invalid => !invalid);
+    if (!isInvalid) {
+      if (ref) {
+        const monthInput = ref.current.querySelector('[name="terra-date-month-date-input-value"]');
+        if (monthInput) {
+          monthInput.focus();
+        }
+      }
+      setErrorMessage('Invalid Departure Date.');
+    } else {
+      setErrorMessage('');
+    }
+  };
+
+  const handleRef = (node) => {
+    ref.current = node;
+  };
 
   return (
     <React.Fragment>
       <DateInputField
-        legend="Legend text"
+        legend="Departure Date"
         name="date-input-value"
         value={value}
         onChange={(event, dateString) => setValue(dateString)}
-        error="Error message"
-        help="Help message"
+        error={errorMessage}
+        help="Note: Departure date should be greater than arrival date"
         isInvalid={isInvalid}
+        refCallback={(node) => handleRef(node)}
       />
       <p>{`DateInputField Value: ${value}`}</p>
-      <button type="button" onClick={() => setIsInvalid(invalid => !invalid)}>Toggle isInvalid</button>
+      <button type="button" onClick={handleClick}>Toggle isInvalid</button>
     </React.Fragment>
   );
 };