+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-input/example/MonthDayYearDateInputField.jsx
@@ -1,10 +1,31 @@
-import React, { useState } from 'react';
+import React, { useState, useRef } from 'react';
 
 import DateInputField from 'terra-date-input/lib/DateInputField';
 
 const Example = () => {
+  const ref = useRef();
   const [value, setValue] = useState('');
   const [isInvalid, setIsInvalid] = useState(false);
+  const [errorMessage, setErrorMessage] = useState('');
+
+  const handleClick = () => {
+    setIsInvalid(invalid => !invalid);
+    if (!isInvalid) {
+      if (ref) {
+        const monthInput = ref.current.querySelector('[name="terra-date-month-date-input-value"]');
+        if (monthInput) {
+          monthInput.focus();
+        }
+      }
+      setErrorMessage('Please enter a valid event date.');
+    } else {
+      setErrorMessage('');
+    }
+  };
+
+  const handleRef = (node) => {
+    ref.current = node;
+  };
 
   return (
     <React.Fragment>
@@ -14,12 +35,13 @@ const Example = () => {
         value={value}
         onChange={(event, dateString) => setValue(dateString)}
         displayFormat="month-day-year"
-        error="Please enter a valid event date."
+        error={errorMessage}
         help="Help message"
         isInvalid={isInvalid}
+        refCallback={(node) => handleRef(node)}
       />
       <p>{`Patient's Birthdate: ${value}`}</p>
-      <button type="button" onClick={() => setIsInvalid(invalid => !invalid)}>Toggle isInvalid</button>
+      <button type="button" onClick={handleClick}>Toggle isInvalid</button>
     </React.Fragment>
   );
 };