+++ packages/terra-date-input/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 1.50.0 - (January 25, 2024)
+
+* Added
+  * Added black dashed border when input gets focused.
+  * Added screen reader support for error messages.
+
 ## 1.49.0 - (January 10, 2024)
 
 * Added