+++ packages/terra-time-input/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+## 4.61.0 - (January 25, 2024)
+
+* Changed
+  * Added metadata as argument for onBlur callback.
+* Fixed
+  * Fixed time input to change the focused input field instead of the minute input while pressing `+`, `-` keys.
+  * Fixed minor formatting mistake.
+
+* Added
+  * Added visual focus dashed border for `terra-time-input`.
+  * Updated screen reader response to announce hot keys suggestions.
+
 ## 4.60.0 - (January 2, 2024)
 
 * Changed